Move AnalyticsManger related classes to the analytics module
COSDK-500
