+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/AnalyticsEvent.kt
@@ -3,10 +3,10 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by ararat on 6/6/2025.
+ * Created by ararat on 27/6/2025.
  */
 
-package com.adyen.checkout.core.internal.analytics
+package com.adyen.checkout.core.analytics.internal
 
 import androidx.annotation.RestrictTo
 import java.util.Date
