+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/DefaultAnalyticsManager.kt
@@ -3,16 +3,14 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by ararat on 6/6/2025.
+ * Created by ararat on 27/6/2025.
  */
 
-package com.adyen.checkout.core.internal.analytics
+package com.adyen.checkout.core.analytics.internal
 
 import androidx.annotation.VisibleForTesting
 import com.adyen.checkout.core.DispatcherProvider
-import com.adyen.checkout.core.internal.analytics.data.AnalyticsRepository
-import com.adyen.checkout.core.analytics.internal.AnalyticsParams
-import com.adyen.checkout.core.analytics.internal.AnalyticsParamsLevel
+import com.adyen.checkout.core.analytics.internal.data.AnalyticsRepository
 import com.adyen.checkout.core.internal.util.runSuspendCatching
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
