+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/ErrorEvent.kt
@@ -3,13 +3,13 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by ararat on 6/6/2025.
+ * Created by ararat on 27/6/2025.
  */
 
-package com.adyen.checkout.core.internal.analytics
+package com.adyen.checkout.core.analytics.internal
 
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.internal.analytics.AnalyticsEvent.Error.Type
+import com.adyen.checkout.core.analytics.internal.AnalyticsEvent.Error.Type
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 enum class ErrorEvent(val errorType: Type, val errorCode: String) {
