+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/data/DefaultAnalyticsRepository.kt
@@ -6,13 +6,13 @@
  * Created by ararat on 6/6/2025.
  */
 
-package com.adyen.checkout.core.internal.analytics.data
+package com.adyen.checkout.core.analytics.internal.data
 
-import com.adyen.checkout.core.internal.analytics.AnalyticsEvent
-import com.adyen.checkout.core.internal.analytics.data.local.AnalyticsLocalDataStore
-import com.adyen.checkout.core.internal.analytics.data.remote.AnalyticsRemoteDataStore
-import com.adyen.checkout.core.internal.analytics.data.remote.AnalyticsSetupProvider
-import com.adyen.checkout.core.internal.analytics.data.remote.AnalyticsTrackRequestProvider
+import com.adyen.checkout.core.analytics.internal.AnalyticsEvent
+import com.adyen.checkout.core.analytics.internal.data.local.AnalyticsLocalDataStore
+import com.adyen.checkout.core.analytics.internal.data.remote.AnalyticsRemoteDataStore
+import com.adyen.checkout.core.analytics.internal.data.remote.AnalyticsSetupProvider
+import com.adyen.checkout.core.analytics.internal.data.remote.AnalyticsTrackRequestProvider
 
 internal class DefaultAnalyticsRepository(
     private val localInfoDataStore: AnalyticsLocalDataStore<AnalyticsEvent.Info>,
