+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/data/local/AnalyticsLocalDataStore.kt
@@ -6,7 +6,7 @@
  * Created by ararat on 6/6/2025.
  */
 
-package com.adyen.checkout.core.internal.analytics.data.local
+package com.adyen.checkout.core.analytics.internal.data.local
 
 internal interface AnalyticsLocalDataStore<T> {
