+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/data/local/InfoAnalyticsLocalDataStore.kt
@@ -6,9 +6,9 @@
  * Created by ararat on 6/6/2025.
  */
 
-package com.adyen.checkout.core.internal.analytics.data.local
+package com.adyen.checkout.core.analytics.internal.data.local
 
-import com.adyen.checkout.core.internal.analytics.AnalyticsEvent
+import com.adyen.checkout.core.analytics.internal.AnalyticsEvent
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import java.util.LinkedList
