+++ core/src/main/java/com/adyen/checkout/core/mbway/internal/ui/MBWayFactory.kt
@@ -11,8 +11,8 @@ package com.adyen.checkout.core.mbway.internal.ui
 import com.adyen.checkout.core.CheckoutConfiguration
 import com.adyen.checkout.core.PaymentMethodTypes
 import com.adyen.checkout.core.internal.PaymentMethodFactory
-import com.adyen.checkout.core.internal.analytics.AnalyticsManagerFactory
-import com.adyen.checkout.core.internal.analytics.AnalyticsSource
+import com.adyen.checkout.core.analytics.internal.AnalyticsManagerFactory
+import com.adyen.checkout.core.analytics.internal.AnalyticsSource
 import com.adyen.checkout.core.internal.ui.model.ButtonComponentParamsMapper
 import com.adyen.checkout.core.internal.ui.model.CommonComponentParamsMapper
 import com.adyen.checkout.core.internal.ui.model.SessionParams
